+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove deduplication based on hashdiffs (#14).
 - Add `CHANGELOG.md` (#16).
 - Make hashkey and hashdiff generation deterministic (#17).
+- Add DISTINCT to filtered_staging in satellite MERGE statement (#18).
 
 ## [0.5.2] - 2021-07-30
 ### Changed